libc: Move com_fd interfaces to comfd.{c,h}
