--- lib/modules/libc/src/fd/comfd_read.c
@@ -1,141 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-#include <assert.h>
-#include <errno.h>
-#include <fcntl.h>
-#include <unistd.h>
-#include <stdlib.h>
-#include <string.h>
-#include <pthread.h>
-#include <stdio.h>
-#include "errcode.h"
-#include "types.h"
-#include "seekop.h"
-#include "rwlock.h"
-#include "counter.h"
-#include "pgtree.h"
-#include "pgtreess.h"
-#include "cmap.h"
-#include "cmapss.h"
-#include "rwlockmap.h"
-#include "rwstatemap.h"
-#include "fcntlop.h"
-#include "ofdid.h"
-#include "ofd.h"
-#include "ofdtx.h"
-#include "fd.h"
-#include "fdtab.h"
-#include "fdtx.h"
-#include "comfd.h"
-
-ssize_t
-com_fd_exec_read(struct com_fd *data, int fildes, void *buf, size_t nbyte, int isnoundo)
-{
-    assert(data);
-
-    /* Update/create fdtx */
-
-    struct fdtx *fdtx = com_fd_get_fdtx(data, fildes);
-    assert(fdtx);
-
-    enum error_code err = fdtx_ref_or_validate(fdtx, fildes, 0);
-
-    if (err) {
-        return err;
-    }
-
-    /* Update/create ofdtx */
-
-    struct ofdtx *ofdtx = com_fd_get_ofdtx(data, fdtx->ofd);
-    assert(ofdtx);
-
-    int optcc;
-    err = ofdtx_ref(ofdtx, fdtx->ofd, fildes, 0, &optcc);
-
-    if (err) {
-        return err;
-    }
-
-    com_fd_set_optcc(data, optcc);
-
-    /* Read */
-
-    enum picotm_libc_validation_mode val_mode = com_fd_get_validation_mode(data);
-
-    int cookie = -1;
-
-    ssize_t len = ofdtx_read_exec(ofdtx,
-                                  fildes, buf, nbyte,
-                                  &cookie, isnoundo, val_mode);
-
-    if (len < 0) {
-        return len;
-    }
-
-    /* possibly validate optimistic domain */
-    if (ofdtx_is_optimistic(ofdtx)
-        && (val_mode == PICOTM_LIBC_VALIDATE_DOMAIN)
-        && ((err = ofdtx_validate(ofdtx)) < 0)) {
-        return err;
-    }
-
-    /* Inject event */
-    if ((cookie >= 0) &&
-        (com_fd_inject(data, ACTION_READ, fildes, cookie) < 0)) {
-        return -1;
-    }
-
-    return len;
-}
-
-int
-com_fd_apply_read(struct com_fd *data, const struct com_fd_event *event, size_t n)
-{
-    assert(data);
-    assert(event || !n);
-
-    int err = 0;
-
-    while (n && !err) {
-
-        assert(event->fildes >= 0);
-        assert(event->fildes < sizeof(data->fdtx)/sizeof(data->fdtx[0]));
-
-        const struct fdtx *fdtx = com_fd_get_fdtx(data, event->fildes);
-        assert(fdtx);
-
-        struct ofdtx *ofdtx = com_fd_get_ofdtx(data, fdtx->ofd);
-        assert(ofdtx);
-
-        int m = 1;
-
-        while ((m < n) && (event[m].fildes == event->fildes)) {
-            ++m;
-        }
-
-        err = ofdtx_read_apply(ofdtx, event->fildes, event, m) < 0 ? -1 : 0;
-        n -= m;
-        event += m;
-    }
-
-    return err;
-}
-
-int
-com_fd_undo_read(struct com_fd *data, int fildes, int cookie)
-{
-    assert(data);
-    assert(fildes >= 0);
-    assert(fildes < sizeof(data->fdtx)/sizeof(data->fdtx[0]));
-
-    const struct fdtx *fdtx = com_fd_get_fdtx(data, fildes);
-    assert(fdtx);
-
-    struct ofdtx *ofdtx = com_fd_get_ofdtx(data, fdtx->ofd);
-    assert(ofdtx);
-
-    return ofdtx_read_undo(ofdtx, fildes, cookie) == (off_t)-1 ? -1 : 0;
-}
-
